+++ src/components/Pagination/Pagination.tsx
@@ -15,14 +15,29 @@ export const Pagination = ({
 }: PaginationProps) => {
   const lastPage = pages <= currentPage + 1
 
+  const navigateTo = (pageIndex: number) => () => {
+    setCurrentPage(pageIndex)
+  }
+
   return (
     <div className="content-pagination">
       <button
         data-fs-button
+        className="pagination__jump-button"
+        onClick={navigateTo(0)}
+        disabled={currentPage === 0}
+      >
+        <HorizontalArrowIcon
+          color={currentPage === 0 ? undefined : '#FF3452'}
+        />
+        <HorizontalArrowIcon
+          color={currentPage === 0 ? undefined : '#FF3452'}
+        />
+      </button>
+      <button
+        data-fs-button
         className="pagination__button"
-        onClick={() => {
-          setCurrentPage(currentPage - 1)
-        }}
+        onClick={navigateTo(currentPage - 1)}
         disabled={currentPage === 0}
       >
         <HorizontalArrowIcon
@@ -41,9 +56,7 @@ export const Pagination = ({
       <h5 className="ofTo">de</h5>
       <button
         className="pagination__last-page"
-        onClick={() => {
-          setCurrentPage(pages - 1)
-        }}
+        onClick={navigateTo(pages - 1)}
         disabled={lastPage}
       >
         {pages}
@@ -51,11 +64,24 @@ export const Pagination = ({
       <button
         data-fs-button
         className="pagination__button"
-        onClick={() => {
-          setCurrentPage(currentPage + 1)
-        }}
+        onClick={navigateTo(currentPage + 1)}
         disabled={lastPage}
       >
+        <HorizontalArrowIcon
+          direction="right"
+          color={lastPage ? undefined : '#FF3452'}
+        />
+      </button>
+      <button
+        data-fs-button
+        className="pagination__jump-button"
+        onClick={navigateTo(pages - 1)}
+        disabled={lastPage}
+      >
+        <HorizontalArrowIcon
+          direction="right"
+          color={lastPage ? undefined : '#FF3452'}
+        />
         <HorizontalArrowIcon
           direction="right"
           color={lastPage ? undefined : '#FF3452'}